backend/master_pipeline/requirements.txt
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
pillow-heif>=0.13.0
insightface>=0.7.3
onnxruntime>=1.16.0
ultralytics>=8.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
tqdm>=4.66.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Optional dependencies
google-auth>=2.23.0
google-api-python-client>=2.100.0
openai>=1.3.0
